Lab updated
Latest error:
Exception in thread "main" java.lang.NullPointerException
at labseven.redblacktree.RBTree.insertFixup(RBTree.java:109)

Tree keeps attempting to access nodes that do not exist. According to
the pseudocode, these nodes should not be a problem, but
NullPointerException is thrown regardless. Multiple attempts to modify
if statements so that they do not execute if parent, parent-parent, etc.
does not yet exist. Perhaps add a height method, and use the method as a
test for the if statements. I.E. if the tree's height is only 2, then
node.getParent().getParent() obviously cannot exist. Fuck this
assignment.